requirements.txt
pypdf
tqdm
numpy
faiss-cpu
openai
sentence-transformers
langgraph
